Standalone: Switch to a hold-and-ask model (part 1)
With these changes, instead of denying execution and then making a rule after authorization, now when in Standalone mode the execution will be allowed but the binary immediately paused with SIGSTOP. If the user authorizes the execution, the binary will be resumed with SIGCONT (as well as a rule being created). While a process is stopped pending user authorization, any other process attempting to send SIGCONT will be blocked. If the GUI is unavailable for any reason, the process will immediately be killed.

This PR does not change logging, so currently all held executions will be logged as if they were an allowed execution even though they may be killed later before the binary ever had the change to execute anything. This will be fixed in a follow-up PR